solutions/02_intermediate/requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
chromadb>=0.4.18
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
tiktoken>=0.5.1
pydantic>=2.4.2
rank_bm25
